+++ projects/ui/src/lib/components/carousel/index.ts
@@ -1,0 +1,2 @@
+export * from './kui-carousel.component';
+export * from './kui-carousel-slide.directive';