+++ projects/ui/src/lib/components/index.ts
@@ -7,6 +7,7 @@ export * from './button';
 export * from './calendar';
 export * from './calendar-range';
 export * from './card';
+export * from './carousel';
 export * from './checkbox';
 export * from './chip';
 export * from './color-input';